fix: common issues
